+++ src/http/index.ts
@@ -1,2 +1,1 @@
 export * from './GraphQLHttpServer';
-export * from './MethodNotAllowedError';